feat- refactor the file structure for components of about related components
