+++ src/components/about/FeatureGrid.tsx
